python/requirements.txt
# Core packages
asyncio
websockets>=10.4
requests>=2.28.0
configparser

# Database
CouchDB>=1.2

# Image processing and ALPR
opencv-python>=4.6.0
numpy>=1.21.0
Pillow>=9.0.0
fast-alpr>=0.1.0

# GPIO (for Raspberry Pi)
RPi.GPIO>=0.7.1

# Audio
pygame>=2.1.0

# Web framework (for admin)
Flask>=2.2.0
Flask-Login>=0.6.0

# Utilities
python-dateutil>=2.8.0
pytz>=2022.1
